refactor: split up use-column-width
