+++ src/pivot-table/components/PivotTable.tsx
@@ -9,7 +9,7 @@ import {
   type ViewService,
 } from "../../types/types";
 import { useStyleContext } from "../contexts/StyleProvider";
-import useColumnWidth from "../hooks/use-column-width";
+import useColumnWidth from "../hooks/use-column-width/use-column-width";
 import useData from "../hooks/use-data";
 import useDataModel from "../hooks/use-data-model";
 import useGridHeight from "../hooks/use-grid-height";
@@ -96,14 +96,14 @@ export const StickyPivotTable = ({
     getRightGridColumnWidth,
     getHeaderCellsIconsVisibilityStatus,
     overrideLeftGridWidth,
-  } = useColumnWidth(
+  } = useColumnWidth({
     layoutService,
     tableRect,
     headersData,
     visibleTopDimensionInfo,
     verticalScrollbarWidth,
     horizontalScrollbarHeightSetter,
-  );
+  });
 
   const { rootWrapper, leftWrapper, rightWrapper } = getScrollableAreasDimensions({
     tableRect,
